consertar erro e atualizar README.md
