--- server/dal/__init__.py
@@ -1,3 +1,0 @@
-from .user import Repository as RepositorioUsuario
-
-
