--- server/domain/__init__.py
@@ -1,1 +1,0 @@
-from .user import getControlador as ControladorUsuario
